requirements.txt
streamlit
Pillow
ultralytics
opencv-python